moved global functions to python file in misc
