+++ mermithid/misc/__init__.py
@@ -6,3 +6,4 @@ from __future__ import absolute_import
 from . import Constants
 from . import KuriePlotTools
 from . import TritiumFormFactor
+from . import FakeTritiumDataFunctions
